@@ -1,21 +1,27 @@
 package jpademo.model;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.Date;
-import java.util.Set;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
 @Table(name = "patient")
+@Entity
 public class Patient {
+
+    //IDENTITY generation type
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "patient_id")
     private Long id;
 
@@ -23,9 +29,18 @@ public class Patient {
     private String name;
 
     @Column(name = "date_of_birth")
-    private Date dateOfBirth;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
+    private LocalDate dateOfBirth;
 
     @OneToMany(mappedBy = "patient",
-    cascade = CascadeType.ALL)
-    private Set<Test> tests;
+            cascade = CascadeType.ALL)
+    private List<Test> tests;
+
+    public void doTest(Test test) {
+        if (tests == null) {
+            tests = new ArrayList<>();
+        }
+        test.setPatient(this);
+        tests.add(test);
+    }
 }
