@@ -1,4 +1,7 @@
 package jpademo.dao;
 
-public interface DeviceDao extends CrudDao<DeviceDao, Long>{
+import jpademo.model.Device;
+
+public interface DeviceDao extends CrudDao<Device, Long> {
+
 }
