@@ -2,5 +2,5 @@ package jpademo.dao;
 
 import jpademo.model.Patient;
 
-public interface PatientDao extends CrudDao<Patient, Long>{
+public interface PatientDao extends CrudDao<Patient, Long> {
 }
