@@ -1,10 +1,11 @@
 package jpademo.util;
 
+import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 
 public class HibernateUtil {
-    private static final String PERSISTENT_UNIT_NAME = "persistence";
+    private static final String PERSISTENT_UNIT_NAME = "persistenceU";
 
     private static final EntityManagerFactory entityManager;
 
@@ -14,6 +15,10 @@ public class HibernateUtil {
         } catch (Throwable ex) {
             throw new ExceptionInInitializerError(ex);
         }
+    }
+
+    public static EntityManager getEntityManager() {
+        return entityManager.createEntityManager();
     }
 
     public static void close() {
